@@ -12,7 +12,7 @@
 		"autoloader-suffix": "Talk",
 		"optimize-autoloader": true,
 		"platform": {
-			"php": "8.1"
+			"php": "8.2"
 		},
 		"sort-packages": true
 	},
