@@ -1,13 +1,13 @@
 {
 	"config": {
 		"platform": {
-			"php": "8.1"
+			"php": "8.2"
 		},
 		"sort-packages": true
 	},
 	"require-dev": {
 		"behat/behat": "^3.25.0",
 		"guzzlehttp/guzzle": "^7.10.0",
-		"phpunit/phpunit": "^10.5.38"
+		"phpunit/phpunit": "^11.5"
 	}
 }
