@@ -1,7 +1,7 @@
 {
 	"config": {
 		"platform": {
-			"php": "8.1"
+			"php": "8.2"
 		}
 	},
 	"require-dev": {
