@@ -1,11 +1,11 @@
 {
 	"config": {
 		"platform": {
-			"php": "8.1"
+			"php": "8.2"
 		},
 		"sort-packages": true
 	},
 	"require-dev": {
-		"phpunit/phpunit": "^10.5.33"
+		"phpunit/phpunit": "^11.5"
 	}
 }
