@@ -1,7 +1,7 @@
 {
 	"config": {
 		"platform": {
-			"php": "8.1.32"
+			"php": "8.2.27"
 		},
 		"sort-packages": true
 	},
